Gen AI-ASSIGNMENTS/ASSIGNMENT-3/requirements.txt
langchain-huggingface
transformers
torch
langchain-core
python-dotenv